Cleanup the required fields cleaner, so it's all in one file
